+++ src/cli/CliMain.h
@@ -1,0 +1,6 @@
+#ifndef CLI_MAIN_H
+#define CLI_MAIN_H
+
+int cliMain(int argc, char* argv[]);
+
+#endif